WebSockets/tests/submissions/Opera/keeping-connection-open/001.html

<!doctype html>
<title>WebSockets: 20s inactivity after handshake</title>
<p>Note: This test takes 20 seconds to run.</p>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_AND_DOMAIN+'/ws/echo');
  ws.onclose = ws.onerror = ws.onmessage = t.step_func(function() {assert_unreached()});
  ws.onopen = function(e) {
    
    setTestTimeout(30000);
    setTimeout(function() {
      ws.send('test');
      ws.onmessage = function(e) {
        
        assert_equals(e.data, 'test');
        ws.onclose = function(e) {
          
          
          setTimeout(t.step_func(function() {t.done();}), 50)
        }
        ws.close();
      }
    }, 20000);
  }
});
</script>
